+++ pdi/src/paraconf_wrapper.cxx
@@ -24,6 +24,8 @@
 
 #include "config.h"
 
+#include <cassert>
+
 #include "pdi/error.h"
 
 #include "pdi/paraconf_wrapper.h"
@@ -42,6 +44,33 @@ void do_pc(PC_tree_t tree, PC_status_t status)
 }
 
 } // namespace
+
+Yaml_region::Yaml_region(PC_tree_t tree)
+	: m_file(
+#ifdef PARACONF_VERSION
+#if PARACONF_UNTYPED_VERSION >= PARACONF_UNTYPED_COMPUTE_VERSION(1, 1, 0)
+		  PC_path(tree)
+#else
+		  ""
+#endif
+#else
+		  ""
+#endif
+	  )
+	, m_start{tree.node->start_mark.line + 1, tree.node->start_mark.column + 1}
+	, m_end{tree.node->end_mark.line + 1, tree.node->end_mark.column + 1}
+{
+	assert(!PC_status(tree) && "building a region from an invalid tree is unspecified");
+}
+
+std::optional<Yaml_region> Yaml_region::make(PC_tree_t tree)
+{
+	if (PC_status(tree) == PC_OK && tree.node) {
+		return Yaml_region(tree);
+	} else {
+		return {};
+	}
+}
 
 Paraconf_wrapper::Paraconf_wrapper()
 	: m_handler{PC_errhandler(PC_NULL_HANDLER)}